tagger/google_vision.py
import io
from google.cloud import vision
from google.cloud.vision import types


def get_image_labels(file_name):
    client = vision.ImageAnnotatorClient()

    with io.open(file_name, 'rb') as image_file:
        content = image_file.read()

    image = types.Image(content=content)

    response = client.label_detection(image=image)
    labels = response.label_annotations

    return labels
