requirements.txt
Django==1.11
MySQL-python==1.2.5
Pillow==4.3.0
django-bootstrap3==9.1.0
google-cloud-vision
